+++ src/common/types/chunk_collection.cpp
@@ -4,6 +4,7 @@
 #include "common/printer.hpp"
 #include "common/value_operations/value_operations.hpp"
 
+#include <algorithm>
 #include <cstring>
 
 using namespace duckdb;
@@ -18,8 +19,8 @@ void ChunkCollection::Append(DataChunk &new_chunk) {
 	// first fill the latest chunk, if it exists
 	count += new_chunk.size();
 
-	size_t remaining_data = new_chunk.size();
-	size_t offset = 0;
+	uint64_t remaining_data = new_chunk.size();
+	uint64_t offset = 0;
 	if (chunks.size() == 0) {
 		// first chunk
 		types = new_chunk.GetTypes();
@@ -28,24 +29,24 @@ void ChunkCollection::Append(DataChunk &new_chunk) {
 		// the types of the new chunk should match the types of the previous one
 		assert(types.size() == new_chunk.column_count);
 		auto new_types = new_chunk.GetTypes();
-		for (size_t i = 0; i < types.size(); i++) {
+		for (uint64_t i = 0; i < types.size(); i++) {
 			assert(new_types[i] == types[i]);
 		}
 #endif
 
 		// first append data to the current chunk
 		DataChunk &last_chunk = *chunks.back();
-		size_t added_data = std::min(remaining_data, STANDARD_VECTOR_SIZE - last_chunk.size());
+		uint64_t added_data = std::min(remaining_data, (uint64_t)(STANDARD_VECTOR_SIZE - last_chunk.size()));
 		if (added_data > 0) {
 			// copy <added_data> elements to the last chunk
-			size_t old_count = new_chunk.size();
-			for (size_t c = 0; c < new_chunk.column_count; c++) {
+			uint64_t old_count = new_chunk.size();
+			for (uint64_t c = 0; c < new_chunk.column_count; c++) {
 				new_chunk.data[c].count = added_data;
 			}
 			last_chunk.Append(new_chunk);
 			remaining_data -= added_data;
 			// reset the chunk to the old data
-			for (size_t c = 0; c < new_chunk.column_count; c++) {
+			for (uint64_t c = 0; c < new_chunk.column_count; c++) {
 				new_chunk.data[c].count = old_count;
 			}
 			offset = added_data;
@@ -67,7 +68,7 @@ void ChunkCollection::Append(DataChunk &new_chunk) {
 // 1 if left > right
 
 template <class TYPE>
-static int8_t templated_compare_value(Vector &left_vec, Vector &right_vec, size_t left_idx, size_t right_idx) {
+static int8_t templated_compare_value(Vector &left_vec, Vector &right_vec, uint64_t left_idx, uint64_t right_idx) {
 	assert(left_vec.type == right_vec.type);
 	TYPE left_val = ((TYPE *)left_vec.data)[left_idx];
 	TYPE right_val = ((TYPE *)right_vec.data)[right_idx];
@@ -80,7 +81,7 @@ static int8_t templated_compare_value(Vector &left_vec, Vector &right_vec, size_
 	return 1;
 }
 
-static int8_t compare_value(Vector &left_vec, Vector &right_vec, size_t vector_idx_left, size_t vector_idx_right) {
+static int8_t compare_value(Vector &left_vec, Vector &right_vec, uint64_t vector_idx_left, uint64_t vector_idx_right) {
 
 	auto left_null = left_vec.nullmask[vector_idx_left];
 	auto right_null = right_vec.nullmask[vector_idx_right];
@@ -115,18 +116,18 @@ static int8_t compare_value(Vector &left_vec, Vector &right_vec, size_t vector_i
 	return false;
 }
 
-static int compare_tuple(ChunkCollection *sort_by, vector<OrderType> &desc, size_t left, size_t right) {
+static int compare_tuple(ChunkCollection *sort_by, vector<OrderType> &desc, uint64_t left, uint64_t right) {
 	assert(sort_by);
 
-	size_t chunk_idx_left = left / STANDARD_VECTOR_SIZE;
-	size_t chunk_idx_right = right / STANDARD_VECTOR_SIZE;
-	size_t vector_idx_left = left % STANDARD_VECTOR_SIZE;
-	size_t vector_idx_right = right % STANDARD_VECTOR_SIZE;
+	uint64_t chunk_idx_left = left / STANDARD_VECTOR_SIZE;
+	uint64_t chunk_idx_right = right / STANDARD_VECTOR_SIZE;
+	uint64_t vector_idx_left = left % STANDARD_VECTOR_SIZE;
+	uint64_t vector_idx_right = right % STANDARD_VECTOR_SIZE;
 
 	auto &left_chunk = sort_by->chunks[chunk_idx_left];
 	auto &right_chunk = sort_by->chunks[chunk_idx_right];
 
-	for (size_t col_idx = 0; col_idx < desc.size(); col_idx++) {
+	for (uint64_t col_idx = 0; col_idx < desc.size(); col_idx++) {
 		auto order_type = desc[col_idx];
 
 		Vector &left_vec = left_chunk->data[col_idx];
@@ -152,7 +153,7 @@ static int64_t _quicksort_initial(ChunkCollection *sort_by, vector<OrderType> &d
 	int64_t pivot = 0;
 	int64_t low = 0, high = sort_by->count - 1;
 	// now insert elements
-	for (size_t i = 1; i < sort_by->count; i++) {
+	for (uint64_t i = 1; i < sort_by->count; i++) {
 		if (compare_tuple(sort_by, desc, i, pivot) <= 0) {
 			result[low++] = i;
 		} else {
@@ -219,9 +220,9 @@ void ChunkCollection::Reorder(uint64_t order_org[]) {
 	auto val_buf = vector<Value>();
 	val_buf.resize(column_count());
 
-	size_t j, k;
-	for (size_t i = 0; i < count; i++) {
-		for (size_t col_idx = 0; col_idx < column_count(); col_idx++) {
+	uint64_t j, k;
+	for (uint64_t i = 0; i < count; i++) {
+		for (uint64_t col_idx = 0; col_idx < column_count(); col_idx++) {
 			val_buf[col_idx] = GetValue(col_idx, i);
 		}
 		j = i;
@@ -231,25 +232,25 @@ void ChunkCollection::Reorder(uint64_t order_org[]) {
 			if (k == i) {
 				break;
 			}
-			for (size_t col_idx = 0; col_idx < column_count(); col_idx++) {
+			for (uint64_t col_idx = 0; col_idx < column_count(); col_idx++) {
 				SetValue(col_idx, j, GetValue(col_idx, k));
 			}
 			j = k;
 		}
-		for (size_t col_idx = 0; col_idx < column_count(); col_idx++) {
+		for (uint64_t col_idx = 0; col_idx < column_count(); col_idx++) {
 			SetValue(col_idx, j, val_buf[col_idx]);
 		}
 	}
 }
 
 template <class TYPE>
-static void templated_set_values(ChunkCollection *src_coll, Vector &tgt_vec, uint64_t order[], size_t col_idx,
-                                 size_t start_offset, size_t remaining_data) {
+static void templated_set_values(ChunkCollection *src_coll, Vector &tgt_vec, uint64_t order[], uint64_t col_idx,
+                                 uint64_t start_offset, uint64_t remaining_data) {
 	assert(src_coll);
 
-	for (size_t row_idx = 0; row_idx < remaining_data; row_idx++) {
-		size_t chunk_idx_src = order[start_offset + row_idx] / STANDARD_VECTOR_SIZE;
-		size_t vector_idx_src = order[start_offset + row_idx] % STANDARD_VECTOR_SIZE;
+	for (uint64_t row_idx = 0; row_idx < remaining_data; row_idx++) {
+		uint64_t chunk_idx_src = order[start_offset + row_idx] / STANDARD_VECTOR_SIZE;
+		uint64_t vector_idx_src = order[start_offset + row_idx] % STANDARD_VECTOR_SIZE;
 
 		auto &src_chunk = src_coll->chunks[chunk_idx_src];
 		Vector &src_vec = src_chunk->data[col_idx];
@@ -263,11 +264,11 @@ static void templated_set_values(ChunkCollection *src_coll, Vector &tgt_vec, uin
 }
 
 // TODO: reorder functionality is similar, perhaps merge
-void ChunkCollection::MaterializeSortedChunk(DataChunk &target, uint64_t order[], size_t start_offset) {
-	size_t remaining_data = min((size_t)STANDARD_VECTOR_SIZE, count - start_offset);
+void ChunkCollection::MaterializeSortedChunk(DataChunk &target, uint64_t order[], uint64_t start_offset) {
+	uint64_t remaining_data = min((uint64_t)STANDARD_VECTOR_SIZE, count - start_offset);
 	assert(target.GetTypes() == types);
 
-	for (size_t col_idx = 0; col_idx < column_count(); col_idx++) {
+	for (uint64_t col_idx = 0; col_idx < column_count(); col_idx++) {
 		target.data[col_idx].count = remaining_data;
 
 		switch (types[col_idx]) {
@@ -300,21 +301,21 @@ void ChunkCollection::MaterializeSortedChunk(DataChunk &target, uint64_t order[]
 	target.Verify();
 }
 
-Value ChunkCollection::GetValue(size_t column, size_t index) {
+Value ChunkCollection::GetValue(uint64_t column, uint64_t index) {
 	return chunks[LocateChunk(index)]->data[column].GetValue(index % STANDARD_VECTOR_SIZE);
 }
 
-vector<Value> ChunkCollection::GetRow(size_t index) {
+vector<Value> ChunkCollection::GetRow(uint64_t index) {
 	vector<Value> values;
 	values.resize(column_count());
 
-	for (size_t p_idx = 0; p_idx < column_count(); p_idx++) {
+	for (uint64_t p_idx = 0; p_idx < column_count(); p_idx++) {
 		values[p_idx] = GetValue(p_idx, index);
 	}
 	return values;
 }
 
-void ChunkCollection::SetValue(size_t column, size_t index, Value value) {
+void ChunkCollection::SetValue(uint64_t column, uint64_t index, Value value) {
 	chunks[LocateChunk(index)]->data[column].SetValue(index % STANDARD_VECTOR_SIZE, value);
 }
 
@@ -333,8 +334,8 @@ bool ChunkCollection::Equals(ChunkCollection &other) {
 		return false;
 	}
 	// if count is equal amount of chunks should be equal
-	for (size_t row_idx = 0; row_idx < count; row_idx++) {
-		for (size_t col_idx = 0; col_idx < column_count(); col_idx++) {
+	for (uint64_t row_idx = 0; row_idx < count; row_idx++) {
+		for (uint64_t col_idx = 0; col_idx < column_count(); col_idx++) {
 			auto lvalue = GetValue(col_idx, row_idx);
 			auto rvalue = other.GetValue(col_idx, row_idx);
 			if (lvalue != rvalue) {
